Add test cases for model Decisions made for ingredient: 1. Ingredients are deemed equal iff both expirydate and name are the same. isSameIngredient was removed. Thus, this is no weaker notion of equal. 2. Quantity now takes in double instead of int. Tests for userPref and GUI are left out

+++ src/test/java/chopchop/testutil/IngredientBuilder.java
@@ -46,7 +46,7 @@ public class IngredientBuilder {
     /**
      * Sets the {@code Address} of the {@code Person} that we are building.
      */
-    public IngredientBuilder withQuantity(int qty) {
+    public IngredientBuilder withQuantity(double qty) {
         this.qty = new Quantity(qty);
         return this;
     }
