+++ src/test/java/chopchop/testutil/IngredientUtil.java
@@ -25,7 +25,7 @@ public class IngredientUtil {
     public static String getIngredientDetails(Ingredient ind) {
         StringBuilder sb = new StringBuilder();
         sb.append(ind.getName().fullName + " ");
-        sb.append(PREFIX_QUANTITY + Integer.toString(ind.getQuantity().value) + " ");
+        sb.append(PREFIX_QUANTITY + Double.toString(ind.getQuantity().value) + " ");
         sb.append(PREFIX_EXPIRY + ind.getExpiryDate().toString() + " ");
         return sb.toString();
     }
